Time: 11 ms (19.92%), Space: 9.8 MB (98.27%) - LeetHub

+++ 350-intersection-of-two-arrays-ii/350-intersection-of-two-arrays-ii.cpp
@@ -3,7 +3,7 @@ public:
     vector<int> intersect(vector<int>& nums1, vector<int>& nums2) {
         sort(nums1.begin(), nums1.end());
         sort(nums2.begin(), nums2.end());
-        int n1 = (int)nums1.size(), n2 = (int)nums2.size();
+        int n1 = nums1.size(), n2 = nums2.size();
         int i1 = 0, i2 = 0;
         vector<int> res;
         while(i1 < n1 && i2 < n2){
